changed drawOrder attribute in CAEntity as float

+++ CrazyArcade/CAFrameWork/CAEntity.cs
@@ -33,8 +33,8 @@ namespace CrazyArcade.CAFramework
         public virtual int X { get => x; set => x = value; }
         public virtual int Y { get => y; set => y = value; }
 
-        protected int DrawOrder = 0;
-        public int ActualDrawOrder => DrawOrder;
+        protected float DrawOrder = 0f;
+        public float ActualDrawOrder => DrawOrder;
 
         public abstract void Load();
         public virtual void Update(GameTime time)
